extract delegate from ConsoleQuentin
