+++ src/main/java/quentin/ConsoleQuentin.java
@@ -1,13 +1,13 @@
 package quentin;
 
 
+import quentin.UI.console.ConsoleGameInitialiser;
 import quentin.UI.console.ConsoleInputHandler;
 import quentin.UI.console.ConsoleOutputHandler;
 import quentin.core.*;
 import quentin.exceptions.QuentinException;
 
 import java.util.InputMismatchException;
-import java.util.NoSuchElementException;
 
 public class ConsoleQuentin extends Quentin<ConsoleInputHandler, ConsoleOutputHandler> {
     public ConsoleQuentin(int boardSize, ConsoleInputHandler inputHandler, ConsoleOutputHandler outputHandler,
@@ -76,47 +76,13 @@ public class ConsoleQuentin extends Quentin<ConsoleInputHandler, ConsoleOutputHa
         }
     }
 
-    private static int getBoardSize(ConsoleInputHandler consoleInputHandler, ConsoleOutputHandler consoleOutputHandler) {
-        consoleOutputHandler.askBoardSize();
-        int boardSize = 0;
-        boolean insertedAValidBoardSize = false;
-
-        while (!insertedAValidBoardSize) {
-            try {
-                boardSize = consoleInputHandler.getInteger();
-                if (boardSize < 4 || boardSize > 13) {
-                    throw new InputMismatchException("Invalid board size! It must be between 4 and 13!");
-                }
-                insertedAValidBoardSize = true;
-            } catch (NoSuchElementException exception) {
-                consoleOutputHandler.notifyException(exception);
-            }
-        }
-        return boardSize;
-    }
-
-    private static Quentin<ConsoleInputHandler, ConsoleOutputHandler> initialise(
-            ConsoleInputHandler inputHandler, ConsoleOutputHandler outputHandler) {
-        outputHandler.displayTitle();
-        outputHandler.displayInstructions();
-        int boardSize = getBoardSize(inputHandler, outputHandler);
-        outputHandler.askBlackPlayerName();
-        String blackPlayerName = inputHandler.askPlayerName();
-        if (blackPlayerName.equals("")) { blackPlayerName = "Player 1"; }
-        outputHandler.askWhitePlayerName();
-        String whitePlayerName = inputHandler.askPlayerName();
-        if (whitePlayerName.equals("")) { whitePlayerName = "Player 2"; }
-        return new ConsoleQuentin(boardSize, inputHandler, outputHandler,
-                blackPlayerName, whitePlayerName);
-    }
-
     public static void main(String... args) {
         boolean wantToReplay = false;
         boolean invalidReplayInput;
         do {
             ConsoleOutputHandler consoleOutputHandler = new ConsoleOutputHandler();
             ConsoleInputHandler consoleInputHandler = new ConsoleInputHandler();
-            initialise(consoleInputHandler, consoleOutputHandler).play();
+            ConsoleGameInitialiser.initialise(consoleInputHandler, consoleOutputHandler).play();
             consoleOutputHandler.printWantToReplay();
             try {
                 wantToReplay = consoleInputHandler.wantToReplay();
